Introduce KeyedDeferredFinalBlockHashAlgorithm<T> and migrate Blake2b/s
Adds KeyedDeferredFinalBlockHashAlgorithm<T> as an abstract base class between
DeferredFinalBlockHashAlgorithm<T> and the Blake2b/Blake2s concrete types.
The new class consolidates the previously duplicated key management code — the
Key property, defensive copying, size validation, InjectKeyBlock(), and secure
disposal of KeyValue — mirroring the existing KeyedBlockHashAlgorithm<T> pattern
for the Merkle-Damgård branch.

Blake2b and Blake2s are refactored to extend the new base; both remove their
local _key field, Key property, OnInitialize() override, Initialize() override,
and InjectKeyBlock() helper, replacing them with a protected override of the
new abstract InitializeHashState() hook.

Adds KeyedDeferredFinalBlockHashAlgorithmTests<TTest, TAlgorithm, TVariant>
as the corresponding abstract test base class, with 17 common Key and
ComputeHash test methods. Blake2bTests and Blake2sTests now inherit from this
class; their Key partial files retain only the algorithm-specific RFC 7693
known-answer tests.
